+++ src/vendus/models/__init__.py
@@ -6,18 +6,21 @@ from vendus.models.client import ClientData
 from vendus.models.document import (
     Document,
     DocumentItem,
+    DocumentMode,
     DocumentStatus,
     DocumentType,
     normalize_status,
 )
-from vendus.models.tax import TaxExemption
+from vendus.models.tax import TaxCategory, TaxExemption
 
 __all__ = [
     "ClientData",
     "Document",
     "DocumentItem",
+    "DocumentMode",
     "DocumentStatus",
     "DocumentType",
+    "TaxCategory",
     "TaxExemption",
     "normalize_status",
 ]